ath6kl: Wait for complete init operation in module_init
This patch make sures that the driver finishes all
the init operation before the control returns to
the user space while doing insmod.

With this patch, there is no delay required in the user
space after loading the driver.  The interface (say, wlan0)
exposed by the driver will be ready and usable immediately
in the user space after loading the driver.

Signed-off-by: Raja Mani <rmani@qca.qualcomm.com>