ath6kl: Fix packet routing issue in AP-AP mode
In AP-AP mode, packets between 2 stations connected on 2 virtual
interfaces are routed with in the driver instead of passing it to
network stack. With bridging enabled, STA1 is connected to
VIF1 and STA2 is connected to VIF2 and started ping between STA1
& STA2 then ping packets were getting routed in ath6kl driver
instead of going to network stack. This issue happens because
connected station list is maintained in driver common data
structure(ar) across VIFs. This patch fixes the issue by
maintaining the VIF pointer in station specific data
structure(ath6kl_sta) and stations are accessed with respect to
VIF to which they are connected.

Signed-off-by: Ram Ranganathan <rrangana@qca.qualcomm.com>